demos/demo-bigbrother/demo-aml/requirements.txt
azureml-core==1.43.0
azureml-pipeline